+++ app/controllers/manager/statistics/articles_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Manager
+  module Statistics
+    class ArticlesController < ApplicationController
+      layout false
+
+      def index; end
+    end
+  end
+end
