--- app/controllers/studies/articles_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Studies
-  class ArticlesController < ApplicationController
-    def index; end
-  end
-end
