+++ app/views/manager/statistics/articles/index.html.erb
