+++ app/views/manager/statistics/employees/index.html.erb
@@ -1,0 +1,12 @@
+<div class="row">
+  <div class="col-sm-6">
+    <div class="card">
+      <div class='card-header'>
+        <%= 'Akademik Personel Sayıları' %>
+      </div>
+      <div class="card-body">
+        <%= column_chart Employee.active.academic.group('titles.name').count, refresh: 1, adapter: :highcharts %>
+      </div>
+    </div>
+  </div>
+</div>
