+++ app/views/manager/statistics/students/cities.html.erb
