+++ app/views/manager/statistics/students/double_major_and_minor.html.erb
