+++ app/views/manager/statistics/students/genders.html.erb
