+++ app/views/manager/statistics/students/genders_and_degrees.html.erb
