+++ app/views/manager/statistics/students/index.html.erb
@@ -1,0 +1,19 @@
+<%= javascript_include_tag 'chart' %>
+
+<div class="row">
+  <div class="col-sm-6 mb-5">
+    <%= content_loader_panel_tag(manager_students_genders_path, refresh: { auto: true }) %>
+  </div>
+  <div class="col-sm-6 mb-5">
+    <%= content_loader_panel_tag(manager_students_genders_and_degrees_path, refresh: { auto: true }) %>
+  </div>
+  <div class="col-sm-6 mb-5">
+    <%= content_loader_panel_tag(manager_students_double_major_and_minor_path, refresh: { auto: true }) %>
+  </div>
+  <div class="col-sm-6 mb-5">
+    <%= content_loader_panel_tag(manager_students_non_graduates_path, refresh: { auto: true }) %>
+  </div>
+  <div class="col-sm-12 mb-5">
+    <%= content_loader_panel_tag(manager_students_cities_path, refresh: { auto: true }) %>
+  </div>
+</div>
