+++ app/views/manager/statistics/students/non_graduates.html.erb
