+++ config/locales/models/studies/en.yml
@@ -1,5 +1,5 @@
 en:
-  studies:
+  manager:
     articles:
       index:
         article: Article
