+++ config/locales/models/studies/tr.yml
@@ -1,17 +1,18 @@
 tr:
-  studies:
-    articles:
-      index:
-        article: Makale
-        number_of_articles_by_years: Yıllara Göre Makale Sayısı
-        publications_by_index_types: Index Türüne Göre Makaleler
-        publications_by_language: Yayın Dili Göre Makaleler
-        publications_by_number_of_authors: Yazar Sayısına Göre Makaleler
-        publications_by_publication_types: Yayın Türüne Göre Makaleler
-        recently_added: En Son Eklenen Makaleler
-        top_publishers: En Çok Makalesi Olanlar
-    dashboard:
-      index:
-        articles: Makaleler
-        certifications: Sertifikalar
-        projects: Projeler
+  manager:
+    statistics:
+      articles:
+        index:
+          article: Makale
+          number_of_articles_by_years: Yıllara Göre Makale Sayısı
+          publications_by_index_types: Index Türüne Göre Makaleler
+          publications_by_language: Yayın Dili Göre Makaleler
+          publications_by_number_of_authors: Yazar Sayısına Göre Makaleler
+          publications_by_publication_types: Yayın Türüne Göre Makaleler
+          recently_added: En Son Eklenen Makaleler
+          top_publishers: En Çok Makalesi Olanlar
+      dashboard:
+        index:
+          articles: Makaleler
+          certifications: Sertifikalar
+          projects: Projeler
