+++ config/routes/manager.rb
@@ -3,11 +3,16 @@
 namespace :manager do
   get '/', to: 'dashboard#index'
 
-  namespace :charts do
-    get :cities
-    get :double_major_and_minor
-    get :genders
-    get :genders_and_degrees
-    get :non_graduates
+  scope module: :statistics do
+    get 'articles', to: 'articles#index'
+    get 'employees', to: 'employees#index'
+    namespace :students do
+      get '/', action: :index
+      get :cities
+      get :double_major_and_minor
+      get :genders
+      get :genders_and_degrees
+      get :non_graduates
+    end
   end
 end
